Rename context and update Loading message
