+++ kinvo-frontend-kleverton-oliveira/src/App.jsx
@@ -2,18 +2,18 @@ import { Box, Flex, Grid, Heading, Spinner } from '@chakra-ui/react';
 import MainContent from './components/MainContent';
 import Navbar from './components/Navbar';
 import Sidebar from './components/Sidebar';
-import { useMainContext } from './context/mainContext';
+import { useApiDataContext } from './context/apiDataContext';
 
 function App() {
 
-  const {isLoading} = useMainContext();
+  const {isLoading} = useApiDataContext();
 
   if(isLoading) {
     return (
       <Grid minH='100vh' bg={'brand.background'} placeItems={'center'}>
         <Flex alignItems={'center'} gap={'2rem'}>
-          <Heading>Loading...</Heading>
-          <Spinner />
+          <Heading color={'brand.darkPurple'}>Loading...</Heading>
+          <Spinner color={'brand.darkPurple'} size='lg'/>
         </Flex>
       </Grid>
     )
