+++ kinvo-frontend-kleverton-oliveira/src/components/MainContent/Portfolio/Portfolio.jsx
@@ -1,15 +1,51 @@
 import { Flex } from '@chakra-ui/react';
 import React from 'react';
-import { useMainContext } from '../../../context/mainContext';
+import { useApiDataContext } from '../../../context/apiDataContext';
 import PortfolioItem from './PortfolioItem';
 
 function Portfolio() {
 
-  const {portfolioData} = useMainContext();
+  const {portfolioData} = useApiDataContext();
+
+  const {
+    equity,
+    equityProfit,
+    indexerValue,
+    percentageOverIndexer,
+    percentageProfit,
+    valueApplied,
+  } = portfolioData;
+
+  const portfolio = [
+    {
+      title: "SALDO BRUTO",
+      value: `R$ ${Number(equity).toLocaleString()}0`,
+    },
+    {
+      title: "VALOR APLICADO",
+      value: `R$ ${Number(valueApplied).toLocaleString()}`,
+    },
+    {
+      title: "RESULTADO",
+      value: `R$ ${Number(equityProfit).toLocaleString()}`,
+    },
+    {
+      title: "RENTABILIDADE",
+      value: `${percentageProfit}%`,
+    },
+    {
+      title: "CDI",
+      value: `${indexerValue}%`,
+    },
+    {
+      title: "% SOBRE CDI",
+      value: `${percentageOverIndexer}%`,
+    },
+  ];
 
   return (
     <Flex mt={'1.5rem'} as='section' gap={'1rem'}>
-      {portfolioData.map((item, index)=>(
+      {portfolio.map((item, index)=>(
           <PortfolioItem key={index} title={item.title} value={item.value}/>
         ))
       }
